api/requirements.txt
fastapi
uvicorn
pydantic
pydantic_ai
sqlalchemy
pgvector
python-dotenv
ruff
mypy
InstructorEmbedding
pypdf
python-docx
